# docs/openapi/bap-v1.yaml
openapi: 3.0.3
info:
  title: Blueprint Agent Protocol (BAP) v1
  version: "2.0.0"
  description: |
    The API surface external autonomous agents (e.g. Hermes) use to operate
    Blueprint without the web dashboard. Mounted at `/api/bap/v1`.

    ## Authentication

    Every route except `POST /register` and `GET /discover` requires a
    `BAP-Key` header (or `Authorization: Bearer <key>`) with a key issued
    by `POST /register`. Keys look like `bap_<36 random base64url chars>`.

    `POST /register` itself requires **either** an authenticated dashboard
    session cookie, **or** a shared `X-Registration-Secret` header matching
    the operator-configured `BAP_REGISTRATION_SECRET` environment variable.
    It is never open to unauthenticated callers.

    Three `GET /agents-admin/*` routes use the dashboard's own session
    cookie instead of `BAP-Key` — they're operator-only and are documented
    separately at the bottom of this spec with the `sessionAuth` scheme.

    ## Idempotency

    Every mutating endpoint (`POST`/`PATCH` that creates or changes state)
    **requires** an `Idempotency-Key` header. Reusing the same key with an
    identical request body replays the original response instead of
    re-running the handler; reusing it with a different body returns `409`.
    A request with no `Idempotency-Key` header on a route that requires one
    returns `400`. See the `IdempotencyKey` parameter and `Conflict`
...

    ## Pagination

...

    ## Timestamps

...

...

...

...

...

  contact:
...
servers:
...
security:
  - bapKey: []

# ─────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bapKey:
      type: apiKey
      in: header
      name: BAP-Key
      description: A key issued by POST /register. Authorization Bearer <key> is also accepted.
    registrationSecret:
      type: apiKey
      in: header
      name: X-Registration-Secret
      description: Shared secret configured via BAP_REGISTRATION_SECRET. Only required if not using a dashboard session.
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Dashboard session cookie — operator-only routes (/agents-admin/*), not part of the external-agent BAP surface.
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string }
      description: Required on every mutating endpoint. See the top-level Idempotency section.
...
    BusinessIdPath:
      name: businessId
      in: path
      required: true
      schema: { type: string }
  responses:
    NotFound:
      description: The resource does not exist.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { error: "Task 'tsk_xxx' not found.", code: not_found, request_id: "..." }
    PermissionDenied:
      description: The caller's key lacks the required permission, or the resource belongs to a business outside the key's business_access.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { error: "Permission denied: task does not belong to an authorized business.", code: permission_denied, request_id: "..." }
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string, enum: [validation_error, not_found, permission_denied, conflict, rate_limited, internal_error] }
        request_id: { type: string, nullable: true }
        retryable: { type: boolean }
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        pages: { type: integer }
...
    Conflict:
      type: object
      properties:
        id: { type: string }
        conflict_type: { type: string }
        severity: { type: string }
        entity_a_type: { type: string }
        entity_a_id: { type: string }
        entity_a_title: { type: string, nullable: true }
        entity_b_type: { type: string }
        entity_b_id: { type: string }
        entity_b_title: { type: string, nullable: true }
        description: { type: string }
        recommendation: { type: string, nullable: true }
        status: { type: string, enum: [open, resolved] }
        detected_at: { type: string, format: date-time }
        resolved_at: { type: string, format: date-time, nullable: true }
...
    Opportunity:
      type: object
      description: "Backed by goal_suggestions (server/brain/goal-suggester.ts), extended in Phase 3."
      properties:
        id: { type: string }
        business_id: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        opportunity_value: { type: number, nullable: true }
        opportunity_unit: { type: string, default: gbp_per_month }
        metric_name: { type: string, nullable: true }
        current_value: { type: number, nullable: true }
        target_value: { type: number, nullable: true }
        barrier: { type: string, nullable: true }
        suggested_deadline: { type: string, format: date-time, nullable: true }
        suggested_agents: { type: array, items: { type: string } }
        suggested_workflow_id: { type: string, nullable: true }
        confidence: { type: number, nullable: true }
        connector_source: { type: string, nullable: true }
        evidence: { type: object }
        status: { type: string, default: pending }
        snoozed_until: { type: string, format: date-time, nullable: true }
        dismissed_reason: { type: string, nullable: true }
        accepted_goal_id: { type: string, nullable: true }
        required_effort: { type: string, nullable: true, description: "Phase 3 addition." }
        related_goal_ids: { type: array, items: { type: string }, description: "Phase 3 addition." }
        related_risks: { type: array, items: { type: object }, description: "Phase 3 addition." }
        why_it_matters: { type: string, nullable: true, description: "Phase 3 addition." }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    GraphNode:
      type: object
      properties:
        id: { type: string, description: "kg_entities.id, an opaque graph node id — not the same as ref_id." }
        entity_type: { type: string, description: "e.g. goal, task, signal, decision, or a standalone type (competitor, product, customer, campaign, person)." }
        label: { type: string }
        ref_table: { type: string, nullable: true, description: "The backing table (goals/tasks/signals/decisions/...), null for standalone entities." }
        ref_id: { type: string, nullable: true }
    GraphEdge:
      type: object
      properties:
        id: { type: string }
        from_entity_id: { type: string }
        to_entity_id: { type: string }
        edge_type:
          type: string
          enum: [caused, supports, blocks, depends_on, relates_to, mentions, supersedes, contradicts, improves, regresses]
        weight: { type: number, default: 1.0 }
        evidence: { type: string, nullable: true }
...
paths:
...
  /businesses/{businessId}/conflicts:
    get:
      operationId: listConflicts
      summary: List conflicts — every conflict_type, filterable, paginated
      x-bap-permission: conflicts:read
      description: >-
        The general, business-wide conflict surface (see also
        GET /goals/{goalId}/conflicts for the goal-scoped view). Read-only —
        resolving or dismissing a conflict remains a dashboard-only action.
      parameters:
        - { $ref: '#/components/parameters/BusinessIdPath' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
        - { name: status, in: query, schema: { type: string, default: open }, description: "'all' for resolved too." }
        - { name: conflict_type, in: query, schema: { type: string }, description: "CSV: goal_vs_goal, task_vs_window, task_vs_goal, task_vs_task, signal_vs_goal, goal_dependency." }
        - { name: category, in: query, schema: { type: string }, description: "CSV: direct, resource, timing, dependency. Phase 3 addition, orthogonal to conflict_type." }
        - { name: entity_id, in: query, schema: { type: string }, description: "Matches either entity_a_id or entity_b_id." }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  conflicts: { type: array, items: { $ref: '#/components/schemas/Conflict' } }
                  total: { type: integer }
                  pagination: { $ref: '#/components/schemas/Pagination' }
  /conflicts/{conflictId}:
    get:
      operationId: getConflictDetail
      summary: Conflict detail
      x-bap-permission: conflicts:read
      parameters: [{ name: conflictId, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  conflict: { $ref: '#/components/schemas/Conflict' }
        '404': { $ref: '#/components/responses/NotFound' }
        '403': { $ref: '#/components/responses/PermissionDenied' }

  # ── Phase 3: Opportunity Engine ─────────────────────────────────────────
  /businesses/{businessId}/opportunities:
    get:
      operationId: listOpportunities
      summary: List quantified opportunities — filterable, paginated
      x-bap-permission: opportunities:read
      parameters:
        - { $ref: '#/components/parameters/BusinessIdPath' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
        - { name: status, in: query, schema: { type: string, default: pending }, description: "CSV, or 'all'." }
        - { name: connector_source, in: query, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  opportunities: { type: array, items: { $ref: '#/components/schemas/Opportunity' } }
                  total: { type: integer }
                  pagination: { $ref: '#/components/schemas/Pagination' }
  /opportunities/{opportunityId}:
    get:
      operationId: getOpportunityDetail
      summary: Opportunity detail
      x-bap-permission: opportunities:read
      parameters: [{ name: opportunityId, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  opportunity: { $ref: '#/components/schemas/Opportunity' }
        '404': { $ref: '#/components/responses/NotFound' }
        '403': { $ref: '#/components/responses/PermissionDenied' }
  /businesses/{businessId}/opportunities/scan:
    post:
      operationId: triggerOpportunityScan
      summary: Trigger a fresh opportunity scan now
      x-bap-permission: opportunities:trigger
      description: >-
        Fire-and-forget — an LLM-backed pattern scan over connector data.
        Poll GET /businesses/{businessId}/opportunities for results.
      parameters:
        - { $ref: '#/components/parameters/BusinessIdPath' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      responses:
        '202': { description: "{ business_id, status: 'scanning', message }" }
        '404': { $ref: '#/components/responses/NotFound' }
        '403': { $ref: '#/components/responses/PermissionDenied' }

  # ── Phase 3: Decision Memory (read-only) ────────────────────────────────
...
  /businesses/{businessId}/graph:
    get:
      operationId: traverseGraph
      summary: Traverse the knowledge graph from a starting entity
      x-bap-permission: graph:read
      parameters:
        - { $ref: '#/components/parameters/BusinessIdPath' }
        - { name: ref_table, in: query, required: true, schema: { type: string, enum: [goals, tasks, signals, decisions, kg_entities] } }
        - { name: ref_id, in: query, required: true, schema: { type: string } }
        - { name: depth, in: query, schema: { type: integer, minimum: 1, maximum: 5, default: 2 } }
        - { name: edge_types, in: query, schema: { type: string }, description: "CSV of caused,supports,blocks,depends_on,relates_to,mentions,supersedes,contradicts,improves,regresses." }
      responses:
        '200':
          description: "OK. { nodes: [], edges: [] } if the starting entity has no graph node yet — not a 404."
          content:
            application/json:
              schema:
                type: object
                properties:
                  nodes: { type: array, items: { $ref: '#/components/schemas/GraphNode' } }
                  edges: { type: array, items: { $ref: '#/components/schemas/GraphEdge' } }
                  total_nodes: { type: integer }
                  total_edges: { type: integer }
        '400': { description: "Missing ref_table/ref_id, or ref_table outside the allowed set." }
        '404': { $ref: '#/components/responses/NotFound' }
  /businesses/{businessId}/graph/rebuild:
    post:
      operationId: rebuildGraph
      summary: Recompute every derivable edge for a business now
      x-bap-permission: graph:trigger
      description: >-
        Synchronous (200, not 202) — idempotent (upserts), safe to call
        repeatedly. Returns total entity/edge counts afterward, not how many
        changed.
      parameters:
        - { $ref: '#/components/parameters/BusinessIdPath' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      responses:
        '200': { description: "{ business_id, entities, edges }" }
        '404': { $ref: '#/components/responses/NotFound' }
        '403': { $ref: '#/components/responses/PermissionDenied' }

  # ── Phase 3: Review surfaces (recommendations, retrospectives, calibration, patterns) ─
...